Move view initialization to GameViewManager
